Switched collection time toggle to a dropdown
